01-MNIST-Digits/requirements.txt
# --- CORE AI STACK ---
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
protobuf==3.20.3
onnx==1.21.0
onnxscript==0.6.2

# --- VISUALIZATION & GUI ---
matplotlib>=3.7.0
Pillow>=9.5.0

# --- TELEMETRY & TOOLS ---
tqdm>=4.65.0
tensorboard>=2.12.0

# --- OPTIONAL: ANALYSIS & EXPORT ---
# onnx>=1.14.0